+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -148,7 +148,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KH3KSVEHREACV1TRJKFXE37H"
+        "01KH3PK052KB719DN9JEBTH2JS"
     )
     r"""The unique identifier of the data source"""