+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KH3KSVEK6GRH8SE8VVTNBZ40"
+        "01KH3PK055Z5CTYEK104J1H124"
     )
     r"""The unique identifier of the data source"""