+++ packages/orq-rc/src/orq_ai_sdk/models/runagentop.py
@@ -2093,7 +2093,7 @@ class RunAgentAgentToolInputRunTools(BaseModel):
 
     schema_: Annotated[AgentToolInputRunSchema, pydantic.Field(alias="schema")]
 
-    id: Optional[str] = "01KH3KSTRZF14YF5CZ6FTVQN76"
+    id: Optional[str] = "01KH3PJZFZWY8FY0BKYE834MVG"
 
     description: Optional[str] = None